requirements.txt
arrow==0.12.1
dateparser==0.7.0
Flask==0.10.1
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
moment==0.8.2
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
regex==2018.8.17
RPi.GPIO==0.6.3
six==1.11.0
SQLAlchemy==1.2.10
times==0.7
tzlocal==1.5.1
Werkzeug==0.14.1
